pyproject.toml
[tool.poetry]
name = "textbase-client"
version = "0.1.2"
description = "A python library to help you create chatbots easily with some more configurations"
authors = ["kaus-cofactory <kaustubhshukla@cofactory.ai>", "sammyCofactory <samarthar@cofactory.ai>"]
readme = "README.md"
packages = [{include = "textbase"}]

[tool.poetry.dependencies]
python = "^3.9.0"
requests = "^2.31.0"
openai = "^0.27.8"
python-dotenv = "^1.0.0"
tabulate = "^0.9.0"
functions-framework = "^3.4.0"
yaspin = "^3.0.0"
pydantic = "^2.3.0"
sounddevice = "0.4.1"
wavio = "0.0.4"
google-cloud-storage = "^2.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
textbase-client = "textbase.textbase_cli:cli"

